WW-15  Fix issues found during code review 3
