+++ WeWatch/WeWatch/HomeView/MovieCardPreviewModel.swift
@@ -1,34 +1,34 @@
-////
-////  DataMovieCardPreviewModel.swift
-////  WeWatch
-////
-////  Created by Anton on 19/01/2025.
-////
 //
-//import Foundation
+//  MovieCardPreviewModel.swift
+//  WeWatch
 //
-//internal struct DataMovieCardPreviewModel: Identifiable, Sendable {
-//    
-//    internal var id: Int
-//    internal var title: String
-//    internal var rating: Double
-//    internal var genres: String
-//    internal var storyline: String
-//    internal var image: URL?
-//    
-//    internal init(
-//        id: Int,
-//        title: String,
-//        rating: Double,
-//        genres: String,
-//        storyline: String,
-//        image: URL?
-//    ) {
-//        self.id = id
-//        self.title = title
-//        self.rating = rating
-//        self.genres = genres
-//        self.storyline = storyline
-//        self.image = image
-//    }
-//}
+//  Created by Anton on 19/01/2025.
+//
+
+import Foundation
+
+internal struct MovieCardPreviewModel: Identifiable, Sendable {
+    
+    internal var id: Int
+    internal var title: String
+    internal var rating: Double
+    internal var genres: String
+    internal var storyline: String
+    internal var image: URL?
+    
+    internal init(
+        id: Int,
+        title: String,
+        rating: Double,
+        genres: String,
+        storyline: String,
+        image: URL?
+    ) {
+        self.id = id
+        self.title = title
+        self.rating = rating
+        self.genres = genres
+        self.storyline = storyline
+        self.image = image
+    }
+}
